requirements.txt
albumentations==1.3.1
cityscapesscripts==2.2.2
easydict==1.10
einops==0.6.1
imageio==2.31.0
ipdb==0.13.13
ipython==8.14.0
matplotlib==3.7.1
mmcv==1.7.0
numpy==1.24.3
opencv_python==4.7.0.72
pandas==2.0.2
Pillow==9.4.0
Pillow==10.0.0
PyYAML==6.0
PyYAML==6.0.1
resampy==0.4.2
scikit_learn==1.2.2
scipy==1.11.2
skimage==0.0
soundfile==0.12.1
tensorboardX==2.6
tensorboardX==2.6.2.2
terminaltables==3.1.10
timm==0.9.5
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.65.0
